+++ apps/src/codebridge/FileBrowser/prompts/openConfirmDeleteFolder.tsx
@@ -6,7 +6,7 @@ import {MultiFileSource} from '@cdo/apps/lab2/types';
 import {
   findFiles,
   findSubFolders,
-} from '@cdo/apps/lab2/utils/multiFileSourceEditUtils';
+} from '@cdo/apps/lab2/utils/multiFileSourceUtils';
 import {DialogType, DialogControlInterface} from '@cdo/apps/lab2/views/dialogs';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
