+++ apps/src/lab2/utils/multiFileSourceEditUtils.ts
@@ -1,45 +1,14 @@
 import {DEFAULT_FOLDER_ID} from '@cdo/apps/codebridge/constants';
 import {getOpenFileIds, sortFilesByName} from '@cdo/apps/codebridge/utils';
 import {getActiveFileForSource} from '@cdo/apps/lab2/projects/utils';
+import {FileId, FolderId, MultiFileSource} from '@cdo/apps/lab2/types';
+
 import {
-  FileId,
-  FolderId,
-  MultiFileSource,
-  ProjectFile,
-  ProjectFolder,
-} from '@cdo/apps/lab2/types';
-
-export const getNextFileId = (files: ProjectFile[]) => {
-  return String(Math.max(0, ...files.map(f => Number(f.id))) + 1);
-};
-
-export const getNextFolderId = (folders: ProjectFolder[]) => {
-  return String(Math.max(0, ...folders.map(f => Number(f.id))) + 1);
-};
-
-export const findSubFolders = (parentId: string, folders: ProjectFolder[]) =>
-  folders.reduce((bucket, f: ProjectFolder) => {
-    if (f.parentId === parentId) {
-      bucket.push(f.id, ...findSubFolders(f.id, folders));
-    }
-    return bucket;
-  }, <string[]>[]);
-
-export const findFiles = (
-  folderId: string,
-  files: ProjectFile[],
-  folders?: ProjectFolder[]
-) => {
-  const folderIds = new Set(
-    folders ? [folderId, ...findSubFolders(folderId, folders)] : [folderId]
-  );
-  return files.reduce((bucket, f: ProjectFile) => {
-    if (folderIds.has(f.folderId)) {
-      bucket.push(f.id);
-    }
-    return bucket;
-  }, <string[]>[]);
-};
+  getNextFileId,
+  getNextFolderId,
+  findSubFolders,
+  findFiles,
+} from './multiFileSourceUtils';
 
 /**
  * Create a new file.
